pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Dot",email = "dot@dotcodes.dev"}
]
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "django (>=5.2.6,<6.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pyproj (>=3.7.2,<4.0.0)",
    "shapely (>=2.1.1,<3.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "whitenoise (>=6.11.0,<7.0.0)",
    "numpy (>=2.3.3,<3.0.0)",
    "django-leaflet (>=0.32.0,<0.33.0)",
    "openai (>=1.108.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
django-extensions = "^4.1"
werkzeug = "^3.1.3"
pyopenssl = "^25.3.0"
django-cors-headers = "^4.9.0"
django-stubs = "^5.2.5"

